+++ astropy/utils/tests/test_misc.py
@@ -23,6 +23,13 @@ def test_isiterable():
     assert misc.isiterable(np.array([1, 2, 3])) is True
 
 
+def test_signal_number_to_name_no_failure():
+    # Regression test that the signal_number_to_name throws no
+    # AttributeError (it tried accessing ".iteritems()" which was removed in
+    # Python3).
+    misc.signal_number_to_name(0)
+
+
 @remote_data
 def test_api_lookup():
     strurl = misc.find_api_page('astropy.utils.misc', 'dev', False, timeout=3)